+++ tests/Language/VisitorTest.php
@@ -49,7 +49,7 @@ final class VisitorTest extends ValidatorTestCase
 
         if ($parent instanceof NodeList) {
             self::assertIsInt($key);
-            self::assertTrue(isset($parent[$key]));
+            self::assertTrue($parent->has($key));
         } else {
             self::assertIsString($key);
             self::assertTrue(property_exists($parent, $key));
@@ -66,7 +66,8 @@ final class VisitorTest extends ValidatorTestCase
         }
 
         if ($parent instanceof NodeList) {
-            self::assertEquals($node, $parent[$key]);
+            self::assertIsInt($key);
+            self::assertEquals($node, $parent->get($key));
         } else {
             /** @phpstan-ignore-next-line */
             self::assertEquals($node, $parent->{$key});
@@ -91,7 +92,8 @@ final class VisitorTest extends ValidatorTestCase
 
         foreach ($path as $key) {
             if ($result instanceof NodeList) {
-                $result = $result[$key];
+                self::assertIsInt($key);
+                $result = $result->get($key);
             } else {
                 /** @phpstan-ignore-next-line */
                 $result = $result->{$key};
@@ -221,7 +223,7 @@ final class VisitorTest extends ValidatorTestCase
         self::assertNotEquals($ast, $editedAst);
 
         $expected = $ast->cloneDeep();
-        $operationNode = $expected->definitions[0];
+        $operationNode = $expected->definitions->get(0);
         assert($operationNode instanceof OperationDefinitionNode);
         $operationNode->directives = new NodeList([$directive1, $directive2]);